VMPSINPOW-175 - List Partner: Adapter, Data class, ViewModel and Mock Data provider

+++ SPApp/app/src/main/java/com/spgroup/spapp/repository/ServicesDataMock.kt
@@ -25,7 +25,7 @@ class ServicesDataMock : ServicesRepository {
                         Supplier(
                         i,
                         "Partner $i",
-                        (165+i) as Float,
+                        165f + i,
                         "month",
                         "testUrl $i",
                         (i % 2 == 0))
